--- src/main/java/com/alibaba/fastjson/serializer/ColorCodec.java
@@ -1,99 +1,0 @@
-package com.alibaba.fastjson.serializer;
-
-import java.awt.Color;
-import java.io.IOException;
-import java.lang.reflect.Type;
-
-import com.alibaba.fastjson.JSON;
-import com.alibaba.fastjson.JSONException;
-import com.alibaba.fastjson.parser.DefaultJSONParser;
-import com.alibaba.fastjson.parser.JSONLexer;
-import com.alibaba.fastjson.parser.JSONToken;
-import com.alibaba.fastjson.parser.deserializer.ObjectDeserializer;
-
-public class ColorCodec implements ObjectSerializer, ObjectDeserializer {
-
-    public final static ColorCodec instance = new ColorCodec();
-
-    public void write(JSONSerializer serializer, Object object, Object fieldName, Type fieldType, int features) throws IOException {
-        SerializeWriter out = serializer.getWriter();
-        Color color = (Color) object;
-        if (color == null) {
-            out.writeNull();
-            return;
-        }
-
-        char sep = '{';
-        if (out.isEnabled(SerializerFeature.WriteClassName)) {
-            out.write('{');
-            out.writeFieldName(JSON.DEFAULT_TYPE_KEY);
-            out.writeString(Color.class.getName());
-            sep = ',';
-        }
-
-        out.writeFieldValue(sep, "r", color.getRed());
-        out.writeFieldValue(',', "g", color.getGreen());
-        out.writeFieldValue(',', "b", color.getBlue());
-        if (color.getAlpha() > 0) {
-            out.writeFieldValue(',', "alpha", color.getAlpha());
-        }
-
-        out.write('}');
-    }
-
-    @SuppressWarnings("unchecked")
-    public <T> T deserialze(DefaultJSONParser parser, Type type, Object fieldName) {
-        JSONLexer lexer = parser.getLexer();
-
-        if (lexer.token() != JSONToken.LBRACE && lexer.token() != JSONToken.COMMA) {
-            throw new JSONException("syntax error");
-        }
-        lexer.nextToken();
-
-        int r = 0, g = 0, b = 0, alpha = 0;
-        for (;;) {
-            if (lexer.token() == JSONToken.RBRACE) {
-                lexer.nextToken();
-                break;
-            }
-
-            String key;
-            if (lexer.token() == JSONToken.LITERAL_STRING) {
-                key = lexer.stringVal();
-                lexer.nextTokenWithColon(JSONToken.LITERAL_INT);
-            } else {
-                throw new JSONException("syntax error");
-            }
-
-            int val;
-            if (lexer.token() == JSONToken.LITERAL_INT) {
-                val = lexer.intValue();
-                lexer.nextToken();
-            } else {
-                throw new JSONException("syntax error");
-            }
-
-            if (key.equalsIgnoreCase("r")) {
-                r = val;
-            } else if (key.equalsIgnoreCase("g")) {
-                g = val;
-            } else if (key.equalsIgnoreCase("b")) {
-                b = val;
-            } else if (key.equalsIgnoreCase("alpha")) {
-                alpha = val;
-            } else {
-                throw new JSONException("syntax error, " + key);
-            }
-
-            if (lexer.token() == JSONToken.COMMA) {
-                lexer.nextToken(JSONToken.LITERAL_STRING);
-            }
-        }
-
-        return (T) new Color(r, g, b, alpha);
-    }
-
-    public int getFastMatchToken() {
-        return JSONToken.LBRACE;
-    }
-}
