+++ src/main/java/com/alibaba/fastjson/serializer/SerializeConfig.java
@@ -227,11 +227,10 @@ public class SerializeConfig extends IdentityHashMap<Type, ObjectSerializer> {
 		// awt
 		if (!awtError) {
     		try {
-    			put(Class.forName("java.awt.Color"), ColorCodec.instance);
-    			put(Class.forName("java.awt.Font"), FontCodec.instance);
-    			put(Class.forName("java.awt.Point"), PointCodec.instance);
-    			put(Class.forName("java.awt.Rectangle"),
-    					RectangleCodec.instance);
+    			put(Class.forName("java.awt.Color"), AwtCodec.instance);
+    			put(Class.forName("java.awt.Font"), AwtCodec.instance);
+    			put(Class.forName("java.awt.Point"), AwtCodec.instance);
+                put(Class.forName("java.awt.Rectangle"), AwtCodec.instance);
     		} catch (Throwable e) {
     		    awtError = true;
     			// skip
