+++ src/test/java/com/alibaba/json/bvt/ColorTest.java
@@ -3,17 +3,18 @@ package com.alibaba.json.bvt;
 import java.awt.Color;
 
 import org.junit.Assert;
-import junit.framework.TestCase;
 
 import com.alibaba.fastjson.JSON;
-import com.alibaba.fastjson.serializer.ColorCodec;
+import com.alibaba.fastjson.serializer.AwtCodec;
 import com.alibaba.fastjson.serializer.JSONSerializer;
+
+import junit.framework.TestCase;
 
 
 public class ColorTest extends TestCase {
     public void test_color() throws Exception {
         JSONSerializer serializer = new JSONSerializer();
-        Assert.assertEquals(ColorCodec.class, serializer.getObjectWriter(Color.class).getClass());
+        Assert.assertEquals(AwtCodec.class, serializer.getObjectWriter(Color.class).getClass());
         
         Color color = Color.RED;
         String text = JSON.toJSONString(color);
