+++ src/test/java/com/alibaba/json/bvt/PointTest.java
@@ -3,17 +3,18 @@ package com.alibaba.json.bvt;
 import java.awt.Point;
 
 import org.junit.Assert;
-import junit.framework.TestCase;
 
 import com.alibaba.fastjson.JSON;
+import com.alibaba.fastjson.serializer.AwtCodec;
 import com.alibaba.fastjson.serializer.JSONSerializer;
-import com.alibaba.fastjson.serializer.PointCodec;
+
+import junit.framework.TestCase;
 
 public class PointTest extends TestCase {
 
     public void test_color() throws Exception {
         JSONSerializer serializer = new JSONSerializer();
-        Assert.assertEquals(PointCodec.class, serializer.getObjectWriter(Point.class).getClass());
+        Assert.assertEquals(AwtCodec.class, serializer.getObjectWriter(Point.class).getClass());
         
         Point point = new Point(3, 4);
         String text = JSON.toJSONString(point);
