+++ src/test/java/com/alibaba/json/bvt/PointTest2.java
@@ -7,14 +7,14 @@ import junit.framework.TestCase;
 
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.serializer.JSONSerializer;
-import com.alibaba.fastjson.serializer.PointCodec;
+import com.alibaba.fastjson.serializer.AwtCodec;
 import com.alibaba.fastjson.serializer.SerializerFeature;
 
 public class PointTest2 extends TestCase {
 
     public void test_point() throws Exception {
         JSONSerializer serializer = new JSONSerializer();
-        Assert.assertEquals(PointCodec.class, serializer.getObjectWriter(Point.class).getClass());
+        Assert.assertEquals(AwtCodec.class, serializer.getObjectWriter(Point.class).getClass());
         
         Point point = new Point(3, 4);
         String text = JSON.toJSONString(point, SerializerFeature.WriteClassName);
