+++ src/test/java/com/alibaba/json/bvt/RectangleTest.java
@@ -3,18 +3,19 @@ package com.alibaba.json.bvt;
 import java.awt.Rectangle;
 
 import org.junit.Assert;
-import junit.framework.TestCase;
 
 import com.alibaba.fastjson.JSON;
+import com.alibaba.fastjson.serializer.AwtCodec;
 import com.alibaba.fastjson.serializer.JSONSerializer;
-import com.alibaba.fastjson.serializer.RectangleCodec;
 import com.alibaba.fastjson.serializer.SerializerFeature;
+
+import junit.framework.TestCase;
 
 public class RectangleTest extends TestCase {
 
     public void test_color() throws Exception {
         JSONSerializer serializer = new JSONSerializer();
-        Assert.assertEquals(RectangleCodec.class, serializer.getObjectWriter(Rectangle.class).getClass());
+        Assert.assertEquals(AwtCodec.class, serializer.getObjectWriter(Rectangle.class).getClass());
         
         Rectangle v = new Rectangle(3, 4, 100, 200);
         String text = JSON.toJSONString(v, SerializerFeature.WriteClassName);
