+++ src/test/java/com/alibaba/json/bvt/parser/deser/FontDeserializerTest.java
@@ -2,18 +2,18 @@ package com.alibaba.json.bvt.parser.deser;
 
 import java.awt.Font;
 
-import junit.framework.TestCase;
-
 import org.junit.Assert;
 
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONException;
-import com.alibaba.fastjson.serializer.FontCodec;
+import com.alibaba.fastjson.serializer.AwtCodec;
+
+import junit.framework.TestCase;
 
 
 public class FontDeserializerTest extends TestCase {
     public void test_0 () throws Exception {
-        FontCodec.instance.getFastMatchToken();
+        AwtCodec.instance.getFastMatchToken();
         
         Assert.assertNull(JSON.parseObject("null", StackTraceElement.class));
         Assert.assertNull(JSON.parseArray("null", StackTraceElement.class));
