+++ src/test/java/com/alibaba/json/bvt/serializer/ColorSerializerTest.java
@@ -2,20 +2,20 @@ package com.alibaba.json.bvt.serializer;
 
 import java.awt.Color;
 
-import junit.framework.TestCase;
-
 import org.junit.Assert;
 
 import com.alibaba.fastjson.JSON;
-import com.alibaba.fastjson.serializer.ColorCodec;
+import com.alibaba.fastjson.serializer.AwtCodec;
 import com.alibaba.fastjson.serializer.JSONSerializer;
 import com.alibaba.fastjson.serializer.SerializerFeature;
+
+import junit.framework.TestCase;
 
 public class ColorSerializerTest extends TestCase {
 
     public void test_null() throws Exception {
         JSONSerializer serializer = new JSONSerializer();
-        Assert.assertEquals(ColorCodec.class, serializer.getObjectWriter(Color.class).getClass());
+        Assert.assertEquals(AwtCodec.class, serializer.getObjectWriter(Color.class).getClass());
 
         VO vo = new VO();
 
@@ -24,7 +24,7 @@ public class ColorSerializerTest extends TestCase {
 
     public void test_rgb() throws Exception {
         JSONSerializer serializer = new JSONSerializer();
-        Assert.assertEquals(ColorCodec.class, serializer.getObjectWriter(Color.class).getClass());
+        Assert.assertEquals(AwtCodec.class, serializer.getObjectWriter(Color.class).getClass());
 
         VO vo = new VO();
         vo.setValue(new Color(1,1,1,0));
