+++ src/test/java/com/alibaba/json/bvt/serializer/FontSerializerTest.java
@@ -3,20 +3,20 @@ package com.alibaba.json.bvt.serializer;
 import java.awt.Font;
 
 import org.junit.Assert;
-import junit.framework.TestCase;
 
 import com.alibaba.fastjson.JSON;
-import com.alibaba.fastjson.serializer.FontCodec;
+import com.alibaba.fastjson.serializer.AwtCodec;
 import com.alibaba.fastjson.serializer.JSONSerializer;
-import com.alibaba.fastjson.serializer.SerializeConfig;
 import com.alibaba.fastjson.serializer.SerializerFeature;
+
+import junit.framework.TestCase;
 
 
 public class FontSerializerTest extends TestCase {
     
     public void test_null() throws Exception {
         JSONSerializer serializer = new JSONSerializer();
-        Assert.assertEquals(FontCodec.class, serializer.getObjectWriter(Font.class).getClass());
+        Assert.assertEquals(AwtCodec.class, serializer.getObjectWriter(Font.class).getClass());
         
         VO vo = new VO();
